requirements.txt
flask==2.0.1
flask-cors==3.0.10
stripe==2.65.0
pyjwt==2.3.0
python-dotenv==0.19.0